fix: wrong api and types files for round api direfctory
